hackyeah2025_backend/requirements.txt
asgiref==3.9.2
Django==5.2.7
psycopg2==2.9.10
sqlparse==0.5.3
Pillow==10.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
